make matchers uniform
* add a few tests
