+++ internal/authorization/subject_matcher.go
@@ -6,23 +6,29 @@ import (
 	"github.com/authelia/authelia/internal/utils"
 )
 
-func isSubjectMatching(subject Subject, subjectRule []string) bool {
-	for _, ruleSubject := range subjectRule {
-		if strings.HasPrefix(ruleSubject, userPrefix) {
-			user := strings.Trim(ruleSubject[len(userPrefix):], " ")
-			if user == subject.Username {
-				continue
-			}
-		}
+func isSubjectMatching(subject Subject, subjectRules [][]string) bool {
+	if len(subjectRules) == 0 {
+		return true
+	}
 
-		if strings.HasPrefix(ruleSubject, groupPrefix) {
-			group := strings.Trim(ruleSubject[len(groupPrefix):], " ")
-			if utils.IsStringInSlice(group, subject.Groups) {
-				continue
+	for _, subjectRule := range subjectRules {
+		for _, ruleSubject := range subjectRule {
+			if strings.HasPrefix(ruleSubject, userPrefix) {
+				user := strings.Trim(ruleSubject[len(userPrefix):], " ")
+				if user == subject.Username {
+					continue
+				}
 			}
-		}
 
-		return false
+			if strings.HasPrefix(ruleSubject, groupPrefix) {
+				group := strings.Trim(ruleSubject[len(groupPrefix):], " ")
+				if utils.IsStringInSlice(group, subject.Groups) {
+					continue
+				}
+			}
+
+			return false
+		}
 	}
 
 	return true
